strategies/cross_exchange_funding_arb/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=5.4.0
